# docs/swagger.yaml
basePath: /
definitions:
  models.BalanceAvailable:
    properties:
      monto_disponible:
        $ref: '#/definitions/models.Moneda'
    type: object
  models.BancoOrigenList:
    properties:
      banco:
        type: string
      id:
        type: integer
      moneda:
        type: string
      nombre:
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      error: {}
    type: object
  models.FacturaDatosBesser:
    properties:
      direccion:
        type: string
      docid:
        type: string
      email:
        type: string
      telefono:
        type: string
    type: object
...
  models.FacturaList:
    properties:
      created_at:
        type: string
      estatus:
        type: string
      factura_id:
        type: string
      monto_pagado:
        $ref: '#/definitions/models.Moneda'
      monto_total:
        $ref: '#/definitions/models.Moneda'
      nreferencia:
        type: string
    type: object
...
  models.FormaPagoList:
    properties:
      banco:
        type: string
      detalle: {}
      id:
        type: integer
      metodo_pago:
        type: string
      moneda:
        type: string
      nombre:
        type: string
    type: object
...
  models.Moneda:
    properties:
      bolivar:
        type: number
      dolar:
        type: number
    type: object
  models.PaginatorData:
    properties:
      current_page:
        type: integer
      next_page:
        type: integer
      prev_page:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
  models.PaymentList:
    properties:
      created_at:
        type: string
      estatus:
        type: string
      fecha:
        type: string
      metodo_pago:
        $ref: '#/definitions/models.FormaPagoList'
      monto_total:
        $ref: '#/definitions/models.Moneda'
      ncontrol:
        type: string
      payment_id:
        type: string
      referencia:
        type: string
    type: object
  models.PaymentReq:
    properties:
      banco_cliente_id:
        maxLength: 15
        minLength: 1
        type: string
      cuenta_banco_id:
        maxLength: 15
        minLength: 1
        type: string
      email:
        type: string
      fecha:
        type: string
      payment_detail:
        items:
          $ref: '#/definitions/models.PaymentReqDetail'
        type: array
      profile_id:
        maxLength: 15
        minLength: 1
        type: string
      referencia:
        maxLength: 200
        minLength: 10
        type: string
      tasa_cambio:
        minimum: 0
        type: number
      telefono:
        type: string
    required:
    - banco_cliente_id
    - cuenta_banco_id
    - fecha
    - payment_detail
    - profile_id
    - referencia
    - tasa_cambio
    type: object
  models.PaymentReqDetail:
    properties:
      monto:
        minimum: 0
        type: number
      suscripcion_id:
        minLength: 1
        type: string
    required:
    - monto
    - suscripcion_id
    type: object
  models.PaymentResponse:
    properties:
      banco_cliente:
        $ref: '#/definitions/models.BancoOrigenList'
      created_at:
        type: string
      datos_besser:
        $ref: '#/definitions/models.FacturaDatosBesser'
      email:
        type: string
      estatus:
        type: string
      facturas:
        items:
          $ref: '#/definitions/models.FacturaList'
        type: array
      fecha:
        type: string
      metodo_pago:
        $ref: '#/definitions/models.FormaPagoList'
      monto_total:
        $ref: '#/definitions/models.Moneda'
      ncontrol:
        type: string
      payment_id:
        type: string
      profile_id:
        type: string
      referencia:
        type: string
      suscripciones:
        items:
          $ref: '#/definitions/models.SuscripcionShortInfo'
        type: array
      tasa_cambio:
        type: number
      telefono:
        type: string
      updated_at:
        type: string
      url_file:
        type: string
    type: object
...
  models.SuccessResponse:
    properties:
      notice:
        type: string
      record: {}
    type: object
  models.SuccessResponseWithMeta:
    properties:
      _meta:
        $ref: '#/definitions/models.PaginatorData'
      notice:
        type: string
      record: {}
    type: object
...
  models.SuscripcionShortInfo:
    properties:
      id:
        type: integer
      monto_pagado:
        $ref: '#/definitions/models.Moneda'
      ncontrol:
        type: string
      speed_unit:
        type: string
      speed_value:
        type: number
      tipo_conexion:
        type: string
      tipo_servicio:
        type: string
      zona:
        type: string
    type: object
  models.TransferList:
    properties:
      created_at:
        type: string
      descripcion:
        type: string
      destinatario_docid:
        type: string
      monto:
        $ref: '#/definitions/models.Moneda'
      ncontrol:
        type: string
      transfer_id:
        type: string
    type: object
  models.TransferReq:
    properties:
      descripcion:
        maxLength: 200
        minLength: 10
        type: string
      destinatario_docid:
        type: string
      monto:
        minimum: 0
        type: number
      profile_id:
        maxLength: 15
        minLength: 1
        type: string
    required:
    - descripcion
    - destinatario_docid
    - monto
    - profile_id
    type: object
  models.TransferResponse:
    properties:
      created_at:
        type: string
      descripcion:
        type: string
      destinatario_docid:
        type: string
      monto:
        $ref: '#/definitions/models.Moneda'
      ncontrol:
        type: string
      transfer_id:
        type: string
    type: object
...
host: 127.0.0.1:7003
info:
  contact:
    email: juan9acosta@gmail.com
    name: Juan Acosta
    url: https://www.linkedin.com/in/juan-m-acosta-f-54219758/
  description: service in Go using Gin framework
  title: MiCuenta Service API
  version: "1.0"
paths:
...
  /payment/image-upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload an image associated with a payment ID
      parameters:
      - description: Access Token
        in: header
        name: x-access-token
        required: true
        type: string
      - description: paymentId (UUID), created_at(timestamptz)
        in: formData
        name: payment_id
        required: true
        type: string
      - description: Image file to upload
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Invalid Request or Incorrect Data
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Upload image for payment
      tags:
      - Payment
  /payment/list:
    get:
      consumes:
      - application/json
      description: Retrieve a list of recibos de pagos with pagination
      parameters:
      - description: Access Token
        in: header
        name: x-access-token
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of records per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponseWithMeta'
            - properties:
                record:
                  items:
                    $ref: '#/definitions/models.PaymentList'
                  type: array
              type: object
        "400":
          description: Invalid Request or Incorrect Data
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Listado de recibos de pago
      tags:
      - Payment
  /payment/send:
    post:
      consumes:
      - application/json
      description: procesa el formulario y valida el mismo, devuelve el id del recibo
        de pago
      parameters:
      - description: Access Token
        in: header
        name: x-access-token
        required: true
        type: string
      - description: Payment Data
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/models.PaymentReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                record:
                  $ref: '#/definitions/models.PaymentResponse'
              type: object
        "400":
          description: Invalid Request or Incorrect Data
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: endpoint para guardar formulario de pago
      tags:
      - Payment
  /payment/show:
    get:
      consumes:
      - application/json
      description: devuelve toda la data relacionada a un recibo de pago
      parameters:
      - description: Access Token
        in: header
        name: x-access-token
        required: true
        type: string
      - description: paymentId (UUID), created_at(timestamptz)
        in: query
        name: PaymentReqId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                record:
                  $ref: '#/definitions/models.PaymentResponse'
              type: object
        "400":
          description: Invalid Request or Incorrect Data
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: detalle de un recibo de pago
      tags:
      - Payment
  /payment/transfer/balance:
    get:
      consumes:
      - application/json
      description: Retrieve the balance available to transfer
      parameters:
      - description: Access Token
        in: header
        name: x-access-token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                record:
                  items:
                    $ref: '#/definitions/models.BalanceAvailable'
                  type: array
              type: object
        "400":
          description: Invalid Request or Incorrect Data
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Saldo disponible para transferir en la cuenta
      tags:
      - Payment
  /payment/transfer/list:
    get:
      consumes:
      - application/json
      description: Retrieve a list of transferencias with pagination
      parameters:
      - description: Access Token
        in: header
        name: x-access-token
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of records per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponseWithMeta'
            - properties:
                record:
                  items:
                    $ref: '#/definitions/models.TransferList'
                  type: array
              type: object
        "400":
          description: Invalid Request or Incorrect Data
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Listado de transferencias
      tags:
      - Payment
  /payment/transfer/send:
    post:
      consumes:
      - application/json
      description: procesa el formulario y valida el mismo, devuelve el id de la transferencia
      parameters:
      - description: Access Token
        in: header
        name: x-access-token
        required: true
        type: string
      - description: Transfer Data
        in: body
        name: transfer
        required: true
        schema:
          $ref: '#/definitions/models.TransferReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                record:
                  $ref: '#/definitions/models.TransferResponse'
              type: object
        "400":
          description: Invalid Request or Incorrect Data
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: endpoint para guardar formulario de transferencia
      tags:
      - Payment
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"